change docker image of ROCm CI to staging_base

+++ fbgemm_gpu/fbgemm_gpu/split_table_batched_embeddings_ops.py
@@ -1977,8 +1977,8 @@ class IntNBitTableBatchedEmbeddingBagsCodegen(nn.Module):
         linear_cache_indices: Tensor,
         offsets: Tensor,
     ) -> None:
-        CACHE_MISS = torch.tensor([-1], device=self.current_device, dtype=linear_cache_indices.dtype)
-        CACHE_HIT = torch.tensor([-2], device=self.current_device, dtype=linear_cache_indices.dtype)
+        CACHE_MISS = torch.tensor([-1], device=self.current_device, dtype=torch.int32)
+        CACHE_HIT = torch.tensor([-2], device=self.current_device, dtype=torch.int32)
 
         # pyre-ignore[6]:
         # Incompatible parameter type [6]: Expected `typing.Sized` for 1st
